refactoring token manager
